Fixed types and updated glide

+++ src/node/core.go
@@ -531,7 +531,7 @@ func (c *Core) GetLastConsensusRoundIndex() *int {
 	return c.poset.LastConsensusRound
 }
 
-func (c *Core) GetConsensusTransactionsCount() int {
+func (c *Core) GetConsensusTransactionsCount() uint64 {
 	return c.poset.ConsensusTransactions
 }
